Split manifest_list into smaller modules
